web/RPS/includes/xssbot/requirements.txt
selenium
pymongo
python-dotenv
webdriver_manager